+++ src/components/Header/Header.css
@@ -1,29 +1,40 @@
 header {
-    background-color: var(--primary);
-    padding: 1rem 2rem;
-  }
-  
+  background-color: var(--primary);
+  padding: 1rem 2rem;
+}
+
+.header-content {
+  max-width: 1200px;
+  margin: 0 auto;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+nav {
+  display: flex;
+  gap: 2rem;
+}
+
+nav a {
+  color: var(--light);
+  text-decoration: none;
+  padding: 0.5rem 1rem;
+  transition: all 0.3s ease;
+}
+
+nav a:hover,
+nav a.active {
+  color: var(--accent);
+}
+
+@media (max-width: 768px) {
   .header-content {
-    max-width: 1200px;
-    margin: 0 auto;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
+    flex-direction: column;
+    gap: 1rem;
   }
   
   nav {
-    display: flex;
-    gap: 2rem;
+    gap: 1rem;
   }
-  
-  nav a {
-    color: var(--light);
-    text-decoration: none;
-    padding: 0.5rem 1rem;
-    transition: all 0.3s ease;
-  }
-  
-  nav a:hover,
-  nav a.active {
-    color: var(--accent);
-  }
+}
