+++ src/pages/About/About.css
@@ -1,0 +1,28 @@
+.about {
+  display: flex;
+  gap: 2rem;
+  padding: 2rem;
+}
+
+.profile-image {
+  width: 300px;
+  height: 300px;
+  border-radius: 50%;
+  object-fit: cover;
+}
+
+.about-content {
+  flex: 1;
+}
+
+.skills ul {
+  list-style: none;
+  padding-left: 1rem;
+}
+
+@media (max-width: 768px) {
+  .about {
+    flex-direction: column;
+    align-items: center;
+  }
+}
