+++ src/pages/About/About.jsx
@@ -1,38 +1,30 @@
-// Import necessary styles
+// Import styles
 import './About.css'
 
 function About() {
   return (
-    // Main section container for About page
     <section className="about">
-      {/* Profile image container */}
+      {/* Profile image section */}
       <div className="profile-image-container">
         <img 
-          src="/path-to-your-image.jpg" // Add your actual image path here
+          src="/path-to-your-image.jpg"  // Update with your image path
           alt="Matthew Chance" 
           className="profile-image"
         />
       </div>
 
-      {/* About content container */}
+      {/* Content section */}
       <div className="about-content">
-        {/* Page title */}
         <h2>About Me</h2>
         
-        {/* Professional introduction */}
         <p className="intro">
           Hello! I'm Matthew Chance, a full-stack web developer passionate about 
-          creating responsive and user-friendly applications. With expertise in React, 
-          JavaScript, and modern web technologies, I focus on building 
-          efficient and scalable solutions.
+          creating responsive and user-friendly applications.
         </p>
 
-        {/* Professional background/skills */}
         <p className="background">
           I specialize in building modern web applications using technologies 
-          like HTML, CSS, JavaScript, React, Node.js, Express. My journey in web development 
-          has equipped me with both front-end and back-end expertise, allowing me 
-          to create comprehensive full-stack solutions.
+          like React, Node.js, Express, and MongoDB.
         </p>
 
         {/* Skills section */}
@@ -50,4 +42,4 @@ function About() {
   )
 }
 
-export default About
+export default About
