+++ src/pages/Portfolio/Portfolio.css
@@ -1,0 +1,112 @@
+.portfolio {
+    padding: 2rem;
+    max-width: 1200px;
+    margin: 0 auto;
+  }
+  
+  .portfolio h2 {
+    color: var(--primary);
+    text-align: center;
+    margin-bottom: 3rem;
+  }
+  
+  /* Projects grid */
+  .projects-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+    gap: 2rem;
+  }
+  
+  /* Project card */
+  .project-card {
+    background: white;
+    border-radius: 8px;
+    overflow: hidden;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease;
+  }
+  
+  .project-card:hover {
+    transform: translateY(-5px);
+  }
+  
+  /* Project image */
+  .project-image {
+    width: 100%;
+    height: 200px;
+    overflow: hidden;
+  }
+  
+  .project-image img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    transition: transform 0.3s ease;
+  }
+  
+  .project-card:hover .project-image img {
+    transform: scale(1.05);
+  }
+  
+  /* Project details */
+  .project-details {
+    padding: 1.5rem;
+  }
+  
+  .project-details h3 {
+    color: var(--primary);
+    margin-bottom: 0.5rem;
+  }
+  
+  .project-details p {
+    color: var(--text);
+    margin-bottom: 1rem;
+  }
+  
+  /* Technologies tags */
+  .technologies {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+    margin-bottom: 1rem;
+  }
+  
+  .tech-tag {
+    background: var(--accent);
+    color: white;
+    padding: 0.25rem 0.75rem;
+    border-radius: 15px;
+    font-size: 0.875rem;
+  }
+  
+  /* Project links */
+  .project-links {
+    display: flex;
+    gap: 1rem;
+  }
+  
+  .project-link {
+    flex: 1;
+    text-align: center;
+    padding: 0.5rem;
+    background: var(--primary);
+    color: white;
+    text-decoration: none;
+    border-radius: 4px;
+    transition: background-color 0.3s ease;
+  }
+  
+  .project-link:hover {
+    background: var(--accent);
+  }
+  
+  /* Responsive design */
+  @media (max-width: 768px) {
+    .portfolio {
+      padding: 1rem;
+    }
+  
+    .projects-grid {
+      grid-template-columns: 1fr;
+    }
+  }
