+++ src/pages/Portfolio/Portfolio.jsx
@@ -106,4 +106,4 @@ function Portfolio() {
     )
 }
 
-export default Portfolio
+export default Portfolio
