+++ src/pages/Portfolio/Resume/Resume.css
@@ -1,0 +1,144 @@
+.resume {
+    max-width: 900px;
+    margin: 0 auto;
+    padding: 2rem;
+  }
+  
+  .resume h2 {
+    color: var(--primary);
+    text-align: center;
+    margin-bottom: 2rem;
+  }
+  
+  /* Download section */
+  .download-section {
+    text-align: center;
+    margin-bottom: 3rem;
+  }
+  
+  .download-btn {
+    background: var(--accent);
+    color: white;
+    padding: 0.75rem 1.5rem;
+    border-radius: 4px;
+    text-decoration: none;
+    transition: background-color 0.3s ease;
+    display: inline-block;
+  }
+  
+  .download-btn:hover {
+    background: var(--accent-dark);
+  }
+  
+  /* Skills section */
+  .skills-section {
+    margin-bottom: 3rem;
+  }
+  
+  .skills-section h3 {
+    color: var(--primary);
+    margin-bottom: 1.5rem;
+  }
+  
+  .skills-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+    gap: 2rem;
+  }
+  
+  .skill-category {
+    background: white;
+    padding: 1.5rem;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .skill-category h4 {
+    color: var(--accent);
+    margin-bottom: 1rem;
+  }
+  
+  .skill-category ul {
+    list-style: none;
+    padding: 0;
+  }
+  
+  .skill-category li {
+    margin-bottom: 0.5rem;
+    padding-left: 1.5rem;
+    position: relative;
+  }
+  
+  .skill-category li::before {
+    content: "•";
+    color: var(--accent);
+    position: absolute;
+    left: 0;
+  }
+  
+  /* Experience section */
+  .experience-section {
+    margin-bottom: 3rem;
+  }
+  
+  .experience-item {
+    margin-bottom: 2rem;
+    padding-bottom: 2rem;
+    border-bottom: 1px solid #eee;
+  }
+  
+  .experience-item h4 {
+    color: var(--primary);
+    margin-bottom: 0.5rem;
+  }
+  
+  .company {
+    color: var(--accent);
+    font-weight: 500;
+    margin-bottom: 0.5rem;
+  }
+  
+  .date {
+    color: var(--secondary);
+    font-size: 0.875rem;
+    margin-bottom: 1rem;
+  }
+  
+  .experience-item ul {
+    padding-left: 1.5rem;
+  }
+  
+  .experience-item li {
+    margin-bottom: 0.5rem;
+  }
+  
+  /* Education section */
+  .education-section {
+    margin-bottom: 2rem;
+  }
+  
+  .education-item {
+    margin-bottom: 2rem;
+  }
+  
+  .education-item h4 {
+    color: var(--primary);
+    margin-bottom: 0.5rem;
+  }
+  
+  .school {
+    color: var(--accent);
+    font-weight: 500;
+    margin-bottom: 0.5rem;
+  }
+  
+  /* Responsive design */
+  @media (max-width: 768px) {
+    .resume {
+      padding: 1rem;
+    }
+  
+    .skills-grid {
+      grid-template-columns: 1fr;
+    }
+  }
